+++ src/condor_ckpt/condor_file_buffer.h
@@ -45,8 +45,8 @@ public:
 
 	virtual int open(const char *url, int flags, int mode);
 	virtual int close();
-	virtual int read(int offset, char *data, int length);
-	virtual int write(int offset, char *data, int length);
+	virtual int read(off_t offset, char *data, int length);
+	virtual int write(off_t offset, char *data, int length);
 
 	virtual int fcntl( int cmd, int arg );
 	virtual int ioctl( int cmd, long arg );
@@ -59,7 +59,7 @@ public:
 	virtual int	is_writeable();
 	virtual int	is_seekable();
 
-	virtual int	get_size();
+	virtual off_t get_size();
 	virtual char const	*get_url();
 
 	virtual int get_unmapped_fd();
@@ -84,7 +84,7 @@ private:
 	int time;
 
 	// The logical size of the buffered file
-	int size;
+	off_t size;
 
 	// The maximum amount of data to buffer
 	int buffer_size;
